+++ aiogram/types/chat_member_member.py
@@ -21,6 +21,8 @@ class ChatMemberMember(ChatMember):
     """The member's status in the chat, always 'member'"""
     user: User
     """Information about the user"""
+    tag: str | None = None
+    """*Optional*. Tag of the member"""
     until_date: DateTime | None = None
     """*Optional*. Date when the user's subscription will expire; Unix time"""
 
@@ -33,6 +35,7 @@ class ChatMemberMember(ChatMember):
             *,
             status: Literal[ChatMemberStatus.MEMBER] = ChatMemberStatus.MEMBER,
             user: User,
+            tag: str | None = None,
             until_date: DateTime | None = None,
             **__pydantic_kwargs: Any,
         ) -> None:
@@ -40,4 +43,6 @@ class ChatMemberMember(ChatMember):
             # This method was auto-generated via `butcher`
             # Is needed only for type checking and IDE support without any additional plugins
 
-            super().__init__(status=status, user=user, until_date=until_date, **__pydantic_kwargs)
+            super().__init__(
+                status=status, user=user, tag=tag, until_date=until_date, **__pydantic_kwargs
+            )