+++ aiogram/types/reply_parameters.py
@@ -24,7 +24,7 @@ class ReplyParameters(TelegramObject):
     allow_sending_without_reply: bool | Default | None = Default("allow_sending_without_reply")
     """*Optional*. Pass :code:`True` if the message should be sent even if the specified message to be replied to is not found. Always :code:`False` for replies in another chat or forum topic. Always :code:`True` for messages sent on behalf of a business account."""
     quote: str | None = None
-    """*Optional*. Quoted part of the message to be replied to; 0-1024 characters after entities parsing. The quote must be an exact substring of the message to be replied to, including *bold*, *italic*, *underline*, *strikethrough*, *spoiler*, and *custom_emoji* entities. The message will fail to send if the quote isn't found in the original message."""
+    """*Optional*. Quoted part of the message to be replied to; 0-1024 characters after entities parsing. The quote must be an exact substring of the message to be replied to, including *bold*, *italic*, *underline*, *strikethrough*, *spoiler*, *custom_emoji*, and *date_time* entities. The message will fail to send if the quote isn't found in the original message."""
     quote_parse_mode: str | Default | None = Default("parse_mode")
     """*Optional*. Mode for parsing entities in the quote. See `formatting options <https://core.telegram.org/bots/api#formatting-options>`_ for more details."""
     quote_entities: list[MessageEntity] | None = None
@@ -33,6 +33,8 @@ class ReplyParameters(TelegramObject):
     """*Optional*. Position of the quote in the original message in UTF-16 code units"""
     checklist_task_id: int | None = None
     """*Optional*. Identifier of the specific checklist task to be replied to"""
+    poll_option_id: str | None = None
+    """*Optional*. Persistent identifier of the specific poll option to be replied to"""
 
     if TYPE_CHECKING:
         # DO NOT EDIT MANUALLY!!!
@@ -51,6 +53,7 @@ class ReplyParameters(TelegramObject):
             quote_entities: list[MessageEntity] | None = None,
             quote_position: int | None = None,
             checklist_task_id: int | None = None,
+            poll_option_id: str | None = None,
             **__pydantic_kwargs: Any,
         ) -> None:
             # DO NOT EDIT MANUALLY!!!
@@ -66,5 +69,6 @@ class ReplyParameters(TelegramObject):
                 quote_entities=quote_entities,
                 quote_position=quote_position,
                 checklist_task_id=checklist_task_id,
+                poll_option_id=poll_option_id,
                 **__pydantic_kwargs,
             )