+++ aiogram/types/user.py
@@ -47,6 +47,8 @@ class User(TelegramObject):
     """*Optional*. :code:`True`, if the bot has forum topic mode enabled in private chats. Returned only in :class:`aiogram.methods.get_me.GetMe`."""
     allows_users_to_create_topics: bool | None = None
     """*Optional*. :code:`True`, if the bot allows users to create and delete topics in private chats. Returned only in :class:`aiogram.methods.get_me.GetMe`."""
+    can_manage_bots: bool | None = None
+    """*Optional*. :code:`True`, if other bots can be created to be controlled by the bot. Returned only in :class:`aiogram.methods.get_me.GetMe`."""
 
     if TYPE_CHECKING:
         # DO NOT EDIT MANUALLY!!!
@@ -70,6 +72,7 @@ class User(TelegramObject):
             has_main_web_app: bool | None = None,
             has_topics_enabled: bool | None = None,
             allows_users_to_create_topics: bool | None = None,
+            can_manage_bots: bool | None = None,
             **__pydantic_kwargs: Any,
         ) -> None:
             # DO NOT EDIT MANUALLY!!!
@@ -92,6 +95,7 @@ class User(TelegramObject):
                 has_main_web_app=has_main_web_app,
                 has_topics_enabled=has_topics_enabled,
                 allows_users_to_create_topics=allows_users_to_create_topics,
+                can_manage_bots=can_manage_bots,
                 **__pydantic_kwargs,
             )