+++ tests/test_filters/test_chat_member_updated.py
@@ -314,6 +314,7 @@ class TestChatMemberUpdatedStatusFilter:
         "can_send_polls": True,
         "can_send_other_messages": True,
         "can_add_web_page_previews": True,
+        "can_edit_tag": True,
         "can_post_stories": True,
         "can_edit_stories": True,
         "can_delete_stories": True,